chore(geofence): post-merge cleanups for location/geofence modules
Three small alignments noticed while diffing the feature branch against main:

- cocoapods_override_sdk.rb: add CustomerIOLocationGeofence to the override
  pod list. Placed under `!is_app_extension` (with Gist/InApp), not the
  "all targets" set — geofencing is a host-app feature (region monitoring +
  launch-time background delivery) and isn't usable in an App Extension.

- Location: revert AppLifecycleNotifying / AppLifecycleObserverToken /
  RealAppLifecycleNotifying from `@_spi(Internal) public` back to internal.
  The geofence module doesn't use them — only Location's own tests do (via
  `@testable import`) — so the SPI exposure was unnecessary. Drops the now-
  pointless `@_spi(Internal)` from GeofenceModuleState's CioLocation import
  and the Location test imports, and shrinks CioLocation.api accordingly.

- Common: move LocationAcquiredEvent into Event.swift (with the other events)
  and LocationData into Type/LocationData.swift (with the other models), then
  delete Event+Location.swift — matching the existing file layout.

https://linear.app/customerio/issue/MBL-1761/ios-execute-geofencing-validation-and-lifecycle-testing
