+++ Tests/Location/LocationServicesImplementationTest.swift
@@ -1,6 +1,6 @@
 @testable import CioInternalCommon
 @testable import CioInternalCommonMocks
-@_spi(Internal) @testable import CioLocation
+@testable import CioLocation
 import CoreLocation
 import SharedTests
 import Testing
