+++ Tests/Location/Stub/NoOpAppLifecycleNotifying.swift
@@ -1,4 +1,4 @@
-@_spi(Internal) @testable import CioLocation
+@testable import CioLocation
 import Foundation
 
 /// No-op implementation of AppLifecycleNotifying for tests that don't need lifecycle behavior.
